front: add the hook useOrderedCriteria.ts
